+++ src/Channel/ExtruderChannel.php
@@ -2,8 +2,12 @@
 
 namespace App\Channel;
 
+use App\ObjectValue\ConnectionLoginData;
+use App\ObjectValue\Message;
+use App\ObjectValue\MessageInterface;
 use App\Service\LoginService;
 use App\Service\ServiceInterface;
+use Closure;
 use Ratchet\ConnectionInterface;
 use Ratchet\MessageComponentInterface;
 use SplObjectStorage;
@@ -15,12 +19,12 @@ use SplObjectStorage;
 class ExtruderChannel implements MessageComponentInterface, ChannelInterface
 {
     use ChannelTrait;
-    
+
     /**
      * @var ConnectionInterface $extruderConnection
      */
     protected $extruderConnection;
-    
+
     /**
      * @var LoginService $loginService
      */
@@ -47,18 +51,25 @@ class ExtruderChannel implements MessageComponentInterface, ChannelInterface
     public function onOpen(ConnectionInterface $conn): void
     {
         $token = $this->getToken($conn->httpRequest);
+
+        /** @var ConnectionLoginData|Message $tokenData */
         $tokenData = $this->loginService->checkLogin($token);
-        
-        if ($tokenData['error']) {
-            $conn->send(json_encode($tokenData));
+
+        if ($tokenData->isError()) {
+            $conn->send(json_encode([
+                "error" => $tokenData->isError(),
+                "message" => $tokenData->getMessage(),
+                "token" => $tokenData->getToken()
+            ]));
             $conn->close();
+            return;
         }
-        
-        if ($tokenData['data']['type'] === 'client') {
+
+        if ($tokenData->isClient()) {
             $this->clients->attach($conn);
         }
-        
-        if ($tokenData['data']['type'] === 'equipament') {
+
+        if ($tokenData->isEquipament()) {
             $this->extruderConnection = $conn;
         }
     }
@@ -68,8 +79,23 @@ class ExtruderChannel implements MessageComponentInterface, ChannelInterface
      */
     public function onClose(ConnectionInterface $conn): void
     {
+        $token = $this->getToken($conn->httpRequest);
+        /** @var ConnectionLoginData $tokenData */
+        $tokenData = $this->loginService->checkLogin($token);
+
         echo "Connection closed\n";
-        $this->clients->detach($conn);
+
+        if ($tokenData->isError()) {
+            return;
+        }
+
+        if ($tokenData->isClient()) {
+            $this->clients->detach($conn);
+        }
+
+        if ($tokenData->isEquipament()) {
+            $this->extruderConnection = null;
+        }
     }
 
     /**
@@ -89,16 +115,59 @@ class ExtruderChannel implements MessageComponentInterface, ChannelInterface
     public function onMessage(ConnectionInterface $conn, $msg): void
     {
         $token = $this->getToken($conn->httpRequest);
+        /** @var MessageInterface|ConnectionLoginData $tokenData */
         $tokenData = $this->loginService->checkLogin($token);
         
-        if ($tokenData['data']['type'] === 'client') {
-            $this->extruderConnection->send($msg);
+        $sendMessage = $this->message($tokenData, $msg);
+
+        $sendMessage($conn);
+    }
+
+    private function message(MessageInterface $tokenData, string $msg): Closure
+    {
+        if ($tokenData->isError()) {
+            return function (ConnectionInterface $conn) use ($tokenData) {
+                $conn->send(json_encode([
+                    'error' => $tokenData->isError(),
+                    'message' => $tokenData->getMessage(),
+                    'token' => $tokenData->getToken()
+                ]));
+                $conn->close();
+            };
         }
 
-        if ($tokenData['data']['type'] === 'equipament') {
-            foreach ($this->clients as $client) {
-                $client->send($msg);
-            }
+        if ($tokenData->isClient() && empty($this->extruderConnection)) {
+            return function (ConnectionInterface $conn) {
+                $conn->send('{"error": true, "message": "Equipament disconected", "token": null}');
+            };
+        }
+
+        if ($tokenData->isEquipament() && empty($this->clients->count())) {
+            return function (ConnectionInterface $conn) {
+                $conn->send('{"error": true, "message": "Client disconected", "token": null}');
+            };
+        }
+
+        if ($tokenData->isClient()) {
+            return function (ConnectionInterface $conn) use ($msg) {
+                $this->extruderConnection->send(json_encode([
+                    'error' => false,
+                    'message' => $msg,
+                    'token' => null
+                ]));
+            };
+        }
+
+        if ($tokenData->isEquipament()) {
+            return function (ConnectionInterface $conn) use ($msg) {
+                foreach ($this->clients as $client) {
+                    $client->send(json_encode([
+                        'error' => false,
+                        'message' => $msg,
+                        'token' => null
+                    ]));
+                }
+            };
         }
     }
 }
